fixed host error

+++ base/views.py
@@ -79,6 +79,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
   room = Room.objects.get(id=pk)
+
+  if request.user != room.host:
+    return HttpResponse('Your are not allowed here!!')
+    
   if request.method == 'POST':
     room.delete()
     return redirect('home')
